Update command line
If there are no phased variants, WhatsHap writes an empty file. Because
of this, we cannot rely on the content of the WhatsHap blocklist files
to determine the sample names.

This commit adds the --whatshap-sample flag to explicitly specify the
sample names, so that all target genes for that sample will appear as
unphased in the MultiQC report.

Also add further checks on invalid combinations of command line
arguments for the target_phasing module.