Ignore command line if --target-genes is missing
This will make it easier to work with this tool in a Cromwell pipeline,
since that doesn't handle optional arrays of files and strings very
well.

Additionally, when there are no target genes specified, it doesn't
matter if --whatshap-sample or --whatshap-blocklist is specified, since
we cannot use those when --target-genes is missing.